refactor: InputUtil Class로 분리
